Fourth update
Drawing of entities in now accessed through an interface - you can now
have entities, which are not for drawing.
Entity class now only handles the actions, copying and data string
saving/reading.
Now it should be easy to add more functionality to the engine by simply
making new interfaces. The next big step in the development is to build
a sound engine.
Game loop is separated from the Plane class - it must be created
manually in the game's code.