Fifth update
Rearranged where classes are written to prepare separate parts of the
code to be extracted and compiled in class libraries for later usage.
Added an image resource database to the plane to allow Drawable entities
to use already loaded images and add more images if needed. Currently it
uses List to store the images, but I will be changing that. Also, you
cannot remove images, this can be as potentially good as it can be bad.
The interface for accessing the images is currently quite bad, but
further usage will give me a better idea on how to set up the whole
thing.