Fifth update
Added some more functionality to the Basis class.
Added algorithms for drawing images: Draw a rotated image, draw a
displaced image, draw image repeatedly as a texture. I have yet to
implement the drawing of a rotated image.
Added an entity database to allow easier handling of the threads of
entities.
Writing XML comments (or however you call it) for documentation. The
documentation text is to be used by Visual Studio to add comments to the
assemblies.